Changed all methods and variable for working with money form double to BigDecimal type

+++ src/main/java/com/github/uinet/dao/imp/DishDAOImp.java
@@ -28,7 +28,7 @@ public class DishDAOImp implements DishDAO {
                 .category(DishCategory.valueOf(resultSet.getString("category")))
                 .name(resultSet.getString("name"))
                 .description(resultSet.getString("description"))
-                .price(resultSet.getDouble("price"))
+                .price(resultSet.getBigDecimal("price"))
                 .image(resultSet.getString("img"))
                 .build();
     }
